requirements.txt
pip==21.3.1
discord.py==1.7.3
PyMySQL==1.0.2
PyNaCl==1.4.0
python-valve==0.2.1
gtts==2.2.3